Moved flag to dedicated flags module.
