+++ rs/registry/canister/src/lib.rs
@@ -4,11 +4,13 @@ pub mod common;
 pub mod get_node_operators_and_dcs_of_node_provider;
 pub mod get_node_providers_monthly_xdr_rewards;
 pub mod init;
-mod invariants;
-mod missing_node_types_map;
 pub mod mutations;
 pub mod pb;
 pub mod proto_on_wire;
 pub mod registry;
 pub mod registry_lifecycle;
 pub mod storage;
+
+mod flags;
+mod invariants;
+mod missing_node_types_map;
