+++ rs/registry/canister/src/storage/tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::flags::temporarily_enable_chunkifying_large_values;
 use ic_registry_transport::pb::v1::{
     high_capacity_registry_mutation, registry_mutation, LargeValueChunkKeys, Precondition,
     RegistryMutation,
